+++ packages/plugin-proxy/index.js
@@ -1,3 +1,0 @@
-const Plugin = require("./src/Plugin");
-
-module.exports = Plugin;