+++ packages/plugin-proxy/src/index.ts
@@ -1,0 +1,6 @@
+import { Plugin } from "./Plugin";
+
+export * from "./types";
+export { Plugin } from "./Plugin";
+
+export default Plugin;